requirements.txt
torch
torchvision
efficientnet-pytorch
pillow
numpy
matplotlib
opencv-python
scipy
